GAS DETECTION ML MODEL + ARDUINO/Implementation 1/GasLeakageDetection/requirements.txt
 
Flask==2.2.2 
requests==2.28.2 
pyserial==3.5 
